created routes separate folder fot routing request used a Route method in express framework
