+++ src/utils/constants.mjs
@@ -1,0 +1,9 @@
+export const mockUsers = [
+    { id: 1, username: "anson", displayName: "Anson" },
+    { id: 2, username: "jack", displayName: "Jack" },
+    { id: 3, username: "adam", displayName: "Adam" },
+    { id: 4, username: "tina", displayName: "Tina" },
+    { id: 5, username: "jason", displayName: "Jason" },
+    { id: 6, username: "henry", displayName: "Henry" },
+    { id: 7, username: "marilyn", displayName: "Marilyn" },
+];
